EMF defaults
