+++ src/Deriver/appsettings.IntegrationTests.json
@@ -1,4 +1,5 @@
 {
+  "AWS_EMF_ENABLED": false,
   "DataApi": {
     "BaseAddress": "http://localhost:9090"
   }
