+++ src/Deriver/appsettings.json
@@ -1,7 +1,5 @@
 {
   "DATA_EVENTS_QUEUE_NAME": "trade_imports_data_upserted_decision_deriver",
-  "AWS_EMF_ENABLED": true,
-  "AWS_EMF_NAMESPACE": "DEFRA/TRADE_IMPORTS/DECISION_DERIVER",
   "Mongo": {
     "DatabaseUri": "mongodb://set-automatically-when-deployed/admin?authSource=$external&authMechanism=MONGODB-AWS",
     "DatabaseName": "trade-imports-decision-deriver"
